requirements.txt
kafka-python==2.0.2
scikit-learn==1.2.2
pandas==2.2.3
joblib==1.3.2
pyarrow==11.0.0
fastparquet==2024.11.0
streamlit==1.27.0
pyspark==3.4.1
pyflink==1.20.0
